WIP: Refaktorere feilmeldinger i vedtak og beregning. Forenklet feilmelding og viser bedre datoformat. Flyttet kode til en annen utils mappe, da den det skal brukes i flere steder i koden.
